requirements.txt
protobuf
groq
cohere
google-genai
python-dotenv
